Reorganizing images and assets and updating imports
