+++ frontend/src/CreateGoal/CreateGoal.css
@@ -1,5 +1,5 @@
 .CreateGoal {
-    background: url("../assets/PinkWave.svg") no-repeat center bottom;
+    background: url("../assets/branding/wave.svg") no-repeat center bottom;
     background-size: cover;
 }
