+++ frontend/src/Home/Home.css
@@ -1,5 +1,5 @@
 .Home {
-    background: url("../assets/wave.svg") no-repeat center bottom;
+    background: url("../assets/branding/wave.svg") no-repeat center bottom;
     min-height: 100vh;
     background-size: cover;
     display: flex;
