+++ frontend/src/Login/Login.css
@@ -1,5 +1,5 @@
 .Login {
-    background: url("../assets/wave.svg") no-repeat center bottom;
+    background: url("../assets/branding/wave.svg") no-repeat center bottom;
     background-size: cover;
 }
