+++ frontend/src/Registration/RegPage.css
@@ -1,6 +1,6 @@
 /* src/RegPage.css */
 .Registration {
-    background: url("../assets/wave.svg") no-repeat center bottom;
+    background: url("../assets/branding/wave.svg") no-repeat center bottom;
     background-size: cover;
 }
